PROG-243 started new auton paths in prep for macomb
