+++ src/main/java/com/lightningrobotics/voidrobot/commands/auto/paths/ThreeBallTerminal.java
@@ -5,16 +5,10 @@ import com.lightningrobotics.common.command.core.TimedCommand;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonDeployIntake;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonIntake;
 import com.lightningrobotics.voidrobot.commands.indexer.AutoIndexCargo;
-import com.lightningrobotics.voidrobot.commands.indexer.RunIndexer;
-import com.lightningrobotics.voidrobot.commands.intake.RunIntake;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonShootCargo;
-import com.lightningrobotics.voidrobot.commands.auto.commands.AutonVisionAim;
-import com.lightningrobotics.voidrobot.commands.shooter.RunShooter;
-import com.lightningrobotics.voidrobot.constants.Constants;
+import com.lightningrobotics.voidrobot.commands.turret.AimTurret;
 import com.lightningrobotics.voidrobot.subsystems.*;
 
-import edu.wpi.first.wpilibj.Timer;
-import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
@@ -23,7 +17,6 @@ public class ThreeBallTerminal extends ParallelCommandGroup {
 
     private static Path start4BallPath = new Path("Start4Ball.path", false);
     private static Path middle4BallPath = new Path("Middle4Ball.path", false);
-    private static Path end4BallPath = new Path("End4Ball.path", false);   
 
 	
     public ThreeBallTerminal(Drivetrain drivetrain, Indexer indexer, Intake intake, Shooter shooter, Hood hood, Turret turret, Vision vision) throws Exception {
@@ -31,19 +24,13 @@ public class ThreeBallTerminal extends ParallelCommandGroup {
 			
 			new TimedCommand(new AutonDeployIntake(intake), 0.75d),
 
-			new AutonVisionAim(vision, turret),
+            new AimTurret(vision, turret, drivetrain),
 
 			new AutoIndexCargo(indexer),
 
 			new SequentialCommandGroup(
-				
-				// Turn On Lights
-				new InstantCommand(vision::turnOnVisionLight, vision),
-
 				// Set Initial Balls Held To 1
 				new InstantCommand(indexer::initializeBallsHeld, indexer),
-    
-				// new TimedCommand(new AutonDeployIntake(intake), 0.75d),
 
 				start4BallPath.getCommand(drivetrain),
 				new AutonShootCargo(shooter, hood, indexer, turret, vision),
