+++ src/main/java/com/lightningrobotics/voidrobot/subsystems/Hood.java
@@ -62,11 +62,11 @@ public class Hood extends SubsystemBase {
 	@Override
 	public void periodic() {	
 		
-		disableHood = hoodDisable.getBoolean(false);
-		if (!manualOverride && !disableHood) {
-			PowerSetPoint = Constants.HOOD_PID.calculate(getAngle(), this.angle);
-			hoodMotor.set(TalonSRXControlMode.PercentOutput, PowerSetPoint);
-		}
+		// disableHood = hoodDisable.getBoolean(false);
+		// if (!manualOverride && !disableHood) {
+		// 	PowerSetPoint = Constants.HOOD_PID.calculate(getAngle(), this.angle);
+		// 	hoodMotor.set(TalonSRXControlMode.PercentOutput, PowerSetPoint);
+		// }
 		
 		setSmartDashboardCommands();
