+++ src/main/java/com/lightningrobotics/voidrobot/subsystems/Indexer.java
@@ -130,6 +130,7 @@ public class Indexer extends SubsystemBase {
         DataLogger.addDataElement("enterSensor", () -> getEnterStatus() ? 1 : 0);
         DataLogger.addDataElement("exitSensor", () -> getExitStatus() ? 1 : 0);
         DataLogger.addDataElement("colorSensor", this::getColorSensorOutputs); // 1 red, 2 blue, 0 nothing 
+        DataLogger.addDataElement("ballCount", this::getBallCount);
     }
 
 	public void initializeBallsHeld() {
